+++ sdk/monitor/azure-monitor-opentelemetry-exporter/azure/monitor/opentelemetry/exporter/_quickpulse/_exporter.py
@@ -1,19 +1,26 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License.
-import logging
-
+from datetime import datetime, timezone
 from enum import Enum
-from typing import Any, Optional
+from typing import Any, Optional, Sequence
 
 from azure.core.exceptions import HttpResponseError
+from azure.monitor.opentelemetry.exporter._quickpulse._constants import _QUICKPULSE_METRIC_NAME_MAPPINGS
 from azure.monitor.opentelemetry.exporter._quickpulse._generated._client import QuickpulseClient
 from azure.monitor.opentelemetry.exporter._quickpulse._generated.models import (
-    CollectionConfigurationInfo,
+    DocumentIngress,
+    MetricPoint,
     MonitoringDataPoint,
 )
 from azure.monitor.opentelemetry.exporter._connection_string_parser import ConnectionStringParser
-from azure.monitor.opentelemetry.exporter._utils import _ticks_since_dot_net_epoch
+from azure.monitor.opentelemetry.exporter._utils import _ticks_since_dot_net_epoch, PeriodicTask
 
+from opentelemetry.context import (
+    _SUPPRESS_INSTRUMENTATION_KEY,
+    attach,
+    detach,
+    set_value,
+)
 from opentelemetry.sdk.metrics import (
     Counter,
     Histogram,
@@ -22,15 +29,18 @@ from opentelemetry.sdk.metrics import (
     ObservableUpDownCounter,
     UpDownCounter,
 )
+from opentelemetry.sdk.metrics._internal.point import (
+    NumberDataPoint,
+    HistogramDataPoint,
+    MetricsData,
+)
 from opentelemetry.sdk.metrics.export import (
     AggregationTemporality,
     MetricExporter,
     MetricExportResult,
     MetricsData as OTMetricsData,
-    PeriodicExportingMetricReader,
+    MetricReader,
 )
-
-_logger = logging.getLogger(__name__)
 
 
 _APPLICATION_INSIGHTS_METRIC_TEMPORALITIES = {
@@ -48,9 +58,17 @@ _LONG_PING_INTERVAL_SECONDS = 60
 _LONG_POST_INTERVAL_SECONDS = 20
 
 
+class Response:
+
+    def __init__(self, pipeline_response, deserialized, response_headers):
+        self._pipeline_response = pipeline_response
+        self._deserialized = deserialized
+        self._response_headers = response_headers
+
+
 class _QuickpulseExporter(MetricExporter):
 
-    def __init__(self, connection_string: str) -> None:
+    def __init__(self, connection_string: Optional[str]) -> None:
         """Metric exporter for Quickpulse.
 
         :param str connection_string: The connection string used for your Application Insights resource.
@@ -58,11 +76,11 @@ class _QuickpulseExporter(MetricExporter):
         """
         parsed_connection_string = ConnectionStringParser(connection_string)
 
-        self._endpoint = parsed_connection_string.endpoint
+        self._live_endpoint = parsed_connection_string.live_endpoint
         self._instrumentation_key = parsed_connection_string.instrumentation_key
         # TODO: Support AADaudience (scope)/credentials
 
-        self._client = QuickpulseClient(host=self._endpoint)
+        self._client = QuickpulseClient(host=self._live_endpoint)
         # TODO: Support redirect
 
         MetricExporter.__init__(
@@ -85,8 +103,35 @@ class _QuickpulseExporter(MetricExporter):
         :return: The result of the export.
         :rtype: ~opentelemetry.sdk.metrics.export.MetricExportResult
         """
-        # TODO
-        return MetricExportResult.SUCCESS
+        result = MetricExportResult.SUCCESS
+        if metrics_data is None:
+            return result
+        data_points = _metric_to_quick_pulse_data_points(
+            metrics_data,
+            base_monitoring_data_point=kwargs.get("base_monitoring_data_point"),
+            documents=kwargs.get("documents"),
+        )
+
+        token = attach(set_value(_SUPPRESS_INSTRUMENTATION_KEY, True))
+        try:
+            post_response = self._client.post(
+                monitoring_data_points=data_points,
+                ikey=self._instrumentation_key,
+                x_ms_qps_transmission_time=_ticks_since_dot_net_epoch(),
+                cls=Response, 
+            )
+            if not post_response:
+                result = MetricExportResult.FAILURE
+            header = post_response._response_headers.get("x-ms-qps-subscribed")
+            if header != "true":
+                # We raise an exception to indicate that quickpulse is not activated anymore
+                raise Exception()
+        except HttpResponseError:
+            # Errors are not reported
+            result = MetricExportResult.FAILURE
+        finally:
+            detach(token)
+        return result
 
     def force_flush(
         self,
@@ -118,21 +163,22 @@ class _QuickpulseExporter(MetricExporter):
         """
 
 
-    def _ping(self, monitoring_data_point) -> Optional[CollectionConfigurationInfo]:
+    def _ping(self, monitoring_data_point) -> Optional[Response]:
+        ping_response = None
+        token = attach(set_value(_SUPPRESS_INSTRUMENTATION_KEY, True))
         try:
             ping_response = self._client.ping(
                 monitoring_data_point=monitoring_data_point,
                 ikey=self._instrumentation_key,
-                x_ms_qps_transmission_time=_ticks_since_dot_net_epoch()
+                x_ms_qps_transmission_time=_ticks_since_dot_net_epoch(),
+                cls=Response, 
             )
-            if isinstance(ping_response, CollectionConfigurationInfo):
-                pass
-            else:
-                # Responses that are not 200s are ignored
-                return None
-        except HttpResponseError as response_error:
+            return ping_response
+        except HttpResponseError:
             # Errors are not reported
-            return None
+            pass
+        detach(token)
+        return ping_response
 
 
 class QuickpulseState(Enum):
@@ -143,10 +189,9 @@ class QuickpulseState(Enum):
     PING_SHORT = _SHORT_PING_INTERVAL_SECONDS
     PING_LONG = _LONG_PING_INTERVAL_SECONDS
     POST_SHORT = _SHORT_POST_INTERVAL_SECONDS
-    POST_LONG = _LONG_POST_INTERVAL_SECONDS
 
 
-class _QuickpulseMetricReader(PeriodicExportingMetricReader):
+class _QuickpulseMetricReader(MetricReader):
 
     def __init__(
         self,
@@ -157,10 +202,17 @@ class _QuickpulseMetricReader(PeriodicExportingMetricReader):
         self._quick_pulse_state = QuickpulseState.PING_SHORT
         self._base_monitoring_data_point = base_monitoring_data_point
         self._elapsed_num_seconds = 0
-        super().__init__(
-            exporter=exporter,
-            export_interval_millis=_SHORT_POST_INTERVAL_SECONDS * 1000,
+        self._worker = PeriodicTask(
+            interval=_SHORT_POST_INTERVAL_SECONDS,
+            function=self._ticker,
+            name="QuickpulseMetricReader",
         )
+        self._worker.daemon = True
+        super().__init__(
+            preferred_temporality=self._exporter._preferred_temporality,
+            preferred_aggregation=self._exporter._preferred_aggregation,
+        )
+        self._worker.start()
 
     def _ticker(self) -> None:
         if self._is_ping_state():
@@ -170,22 +222,95 @@ class _QuickpulseMetricReader(PeriodicExportingMetricReader):
                 ping_response = self._exporter._ping(
                     self._base_monitoring_data_point,
                 )
-                if ping_response and ping_response.response_headers:
-                    if ping_response.response_headers.get("x-ms-qps-subscribed"):
+                if ping_response:
+                    header = ping_response._response_headers.get("x-ms-qps-subscribed")
+                    if header and header == "true":
+                        print("ping succeeded: switching to post")
                         # Switch state to post if subscribed
                         self._quick_pulse_state = QuickpulseState.POST_SHORT
+                        self._elapsed_num_seconds = 0
+                    else:
+                        # Backoff after _LONG_PING_INTERVAL_SECONDS (60s) of no successful requests
+                        if self._quick_pulse_state is QuickpulseState.PING_SHORT and self._elapsed_num_seconds >= _LONG_PING_INTERVAL_SECONDS:
+                            print("ping failed for 60s, switching to pinging every 60s")
+                            self._quick_pulse_state = QuickpulseState.PING_LONG
                 # TODO: Implement redirect
-                # TODO: Implement interval hint
                 else:
                     # Erroroneous responses instigate backoff logic
                     # Backoff after _LONG_PING_INTERVAL_SECONDS (60s) of no successful requests
-                    if self._elapsed_num_seconds >= _LONG_PING_INTERVAL_SECONDS:
+                    if self._quick_pulse_state is QuickpulseState.PING_SHORT and self._elapsed_num_seconds >= _LONG_PING_INTERVAL_SECONDS:
+                        print("ping failed for 60s, switching to pinging every 60s")
                         self._quick_pulse_state = QuickpulseState.PING_LONG
-                    pass
         else:
-            print("posting")
-            pass
+            print("posting...")
+            try:
+                self.collect()
+            except Exception as ex:  # pylint: disable=broad-except,invalid-name
+                # Unsuccessful exports instigate backoff logic
+                # Backoff after _LONG_POST_INTERVAL_SECONDS (20s) of no successful requests
+                # And resume pinging
+                if self._elapsed_num_seconds >= _LONG_POST_INTERVAL_SECONDS:
+                    print("post failed for 20s, switching to pinging")
+                    self._quick_pulse_state = QuickpulseState.PING_SHORT
+                    self._elapsed_num_seconds = 0
+
+        self._elapsed_num_seconds += 1
+
+    def _receive_metrics(
+        self,
+        metrics_data: MetricsData,
+        timeout_millis: float = 10_000,
+        **kwargs,
+    ) -> None:
+        result = self._exporter.export(
+            metrics_data,
+            timeout_millis=timeout_millis,
+            base_monitoring_data_point=self._base_monitoring_data_point,
+            documents=[],
+        )
+        if result is MetricExportResult.FAILURE:
+            # There is currently no way to propagate unsuccessful metric post so we raise an exception
+            # MUST handle this exception whenever `collect()` is called
+            raise Exception()
+
+    def shutdown(self, timeout_millis: float = 30_000, **kwargs) -> None:
+        self._worker.cancel()
+        self._worker.join()
 
     def _is_ping_state(self):
         return self._quick_pulse_state in (QuickpulseState.PING_SHORT, QuickpulseState.PING_LONG)
     
+def _metric_to_quick_pulse_data_points(
+    metrics_data: OTMetricsData,
+    base_monitoring_data_point: MonitoringDataPoint,
+    documents: Sequence[DocumentIngress],
+) -> Sequence[MonitoringDataPoint]:
+    metric_points = []
+    for resource_metric in metrics_data.resource_metrics:
+        for scope_metric in resource_metric.scope_metrics:
+            for metric in scope_metric.metrics:
+                for point in metric.data.data_points:
+                    if point is not None:
+                        metric_point = MetricPoint(
+                            name=_QUICKPULSE_METRIC_NAME_MAPPINGS[metric.name.lower()],
+                            weight=1,
+                        )
+                        if isinstance(point, HistogramDataPoint):
+                            metric_point.value = point.sum
+                        elif isinstance(point, NumberDataPoint):
+                            metric_point.value = point.value
+                        else:
+                            metric_point.value = 0
+                        metric_points.append(metric_point)
+    return [
+        MonitoringDataPoint(
+            version=base_monitoring_data_point.version,
+            instance=base_monitoring_data_point.instance,
+            role_name=base_monitoring_data_point.role_name,
+            machine_name=base_monitoring_data_point.machine_name,
+            stream_id=base_monitoring_data_point.stream_id,
+            timestamp=datetime.now(tz=timezone.utc),
+            metrics=metric_points,
+            documents=documents,
+        )
+    ]