+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Add live metrics skeleton + swagger definitions
     ([#33983](https://github.com/Azure/azure-sdk-for-python/pull/33983))
+- Add live metrics exporting functionality
+    ([#33983](https://github.com/Azure/azure-sdk-for-python/pull/33983))
 
 ### Breaking Changes